agentp/requirements.txt
livekit-agents~=0.11.0
livekit-plugins-openai~=0.10.5
livekit-plugins-deepgram>=0.6.9
livekit-plugins-rag>=0.2.2
livekit-plugins-silero~=0.7.3
python-dotenv~=1.0
google-api-python-client~=2.151.0
google-auth-oauthlib~=1.2.1
google-auth-httplib2~=0.2.0
